+++ mColors.py
@@ -1,9 +1,13 @@
 from oConsole import oConsole;
 
 # Colors used in output for various types of information:
-NORMAL =  0x0F07;  # Console default color
-INFO =    0x0F0A;  # Light green (foreground only)
-HILITE =  0x0F0F;  # White (foreground only)
-ERROR =   0x0F0C;  # Light red (foreground only)
+NORMAL =      0x0F07;  # White / Light grey
+DIM =         0x0F08;  # Dark gray / bright black
+HEADER =      0xFF8F;  # Bright White on Dark gray / bright black
+SELECTED =    0xF010;  # Dark blue background
+INFO =        0x0F0A;  # Bright green
+HILITE =      0x0F0F;  # Bright White
+ERROR =       0x0F04;  # Red
+ERROR_INFO =  0x0F0C;  # Bright red
 oConsole.uDefaultColor = NORMAL;